+++ src/edu/ucsb/cs56/F16/CoursesScraper/CourseScraperTest.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.W15.GEscraper;
+package edu.ucsb.cs56.F16.CourseScraper;
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import java.util.*;
@@ -6,38 +6,38 @@ import java.io.*;
 import java.net.*;
 
 
-/** test class for GEscraper
+/** test class for CoursesScraper
 
 @author Brian Wan
 @author Dylan Lynch
 @version 03.04.15
-@see GEscraper
+@see CoursesScraper
 
 */
 
-public class GEscraperTest {
+public class CourseScraperTest {
 
     /** Test getSpecificCourses with General Course Input
-    @see GEscraper
+    @see CoursesScraper
     */
     @Test
     public void test_getSpecificCourses() {
-      GEscraper g = new GEscraper();
-      // ArrayList with only Area B French courses
-      ArrayList<String> f = g.getSpecificCourses("B", "FR");
-      // ArrayList with all Area B Courses
-      ArrayList<String> a = g.getCourses("B");
-      System.out.println("getCourses:         " + a.get(7));
-      System.out.println("getSpecificCourses: " + f.get(0));
-      assertEquals(f.get(0), a.get(7));
+      // GetInfo g = new GetInfo();
+      // // ArrayList with only Area B French courses
+      // ArrayList<String> f = g.getSpecificCourses("B", "FR");
+      // // ArrayList with all Area B Courses
+      // ArrayList<String> a = g.getCourses("B");
+      // System.out.println("getCourses:         " + a.get(7));
+      // System.out.println("getSpecificCourses: " + f.get(0));
+      // assertEquals(f.get(0), a.get(7));
     }
 
     // /** Test getCourses with General Course Input
-    // @see GEscraper
+    // @see CoursesScraper
     // */
     // @Test
     // public void test_getCoursesGeneral(){
-    //   GEscraper g = new GEscraper();
+    //   CoursesScraper g = new CoursesScraper();
     //   ArrayList<String> s = g.getCourses("B");
 
     //   String a = s.get(0);
@@ -45,11 +45,11 @@ public class GEscraperTest {
     // }
 
     // * Test getCourses with Special Course Subject Input"
-    // @see GEscraper
+    // @see CoursesScraper
     
     // @Test
     // public void test_getCoursesSpecial(){
-    //   GEscraper g = new GEscraper();
+    //   CoursesScraper g = new CoursesScraper();
     //   ArrayList<String> s = g.getCourses("QNT");
 
     //   String a = s.get(0);
